move styling for the filter control component into the component itself
